GH-111: create s3 catalog component
