+++ .cargo/audit.toml
@@ -4,9 +4,9 @@ ignore = [
     "RUSTSEC-2025-0134", # `rustls-pemfile`, is unmaintained
     # bincode 1.x is pulled in transitively via foyer/iceberg; this advisory is tracked as an unmaintained dependency warning, not a known vulnerability in our direct code.
     "RUSTSEC-2025-0141",
-    # TODO: rustls-webpki CRL panic; transitive via aws-smithy-http-client -> rustls 0.21 (0.101.7) and testcontainers -> reqwest -> rustls-platform-verifier (0.103.12). Remove once upstream drops legacy rustls.
+    # TODO: rustls-webpki 0.101.7 CRL matching flaw; transitive via aws-smithy-http-client -> rustls 0.21. Remove once AWS SDK drops legacy rustls.
     "RUSTSEC-2026-0049",
-    "RUSTSEC-2026-0104",
     "RUSTSEC-2026-0098",
-    "RUSTSEC-2026-0099"
+    "RUSTSEC-2026-0099",
+    "RUSTSEC-2026-0104"
 ]
